refactor: move ApiToken type to token.types.ts
