+++ packages/lib/modules/marketing/useEcosystemPoolActivity.tsx
@@ -17,7 +17,7 @@ import { NumberFormatter } from '@repo/lib/shared/utils/numbers'
 import { usePoolEvents } from '../pool/usePoolEvents'
 import { supportedNetworks } from '../web3/ChainConfig'
 import { getChainShortName } from '@repo/lib/config/app.config'
-import { ApiToken } from '../pool/pool.types'
+import { ApiToken } from '../tokens/token.types'
 import {
   getBlockExplorerAddressUrl,
   getBlockExplorerTxUrl,
