+++ packages/lib/modules/pool/PoolDetail/PoolInfo/Erc4626Info.tsx
@@ -17,7 +17,7 @@ import Link from 'next/link'
 import { ArrowUpRight } from 'react-feather'
 import { getWarnings } from '../../pool.helpers'
 import { PropsWithChildren } from 'react'
-import { ApiToken } from '../../pool.types'
+import { ApiToken } from '@repo/lib/modules/tokens/token.types'
 
 type Erc4626InfoPopOverProps = {
   token: ApiToken
