+++ packages/lib/modules/tokens/TokenRow/TokenRowGroup.tsx
@@ -1,7 +1,7 @@
 import { HStack, Skeleton, Text, VStack } from '@chakra-ui/react'
 import { useCurrency } from '@repo/lib/shared/hooks/useCurrency'
 import { GqlChain } from '@repo/lib/shared/services/api/generated/graphql'
-import { ApiToken } from '../../pool/pool.types'
+import { ApiToken } from '../token.types'
 import { HumanTokenAmount } from '../token.types'
 import { useTotalUsdValue } from '../useTotalUsdValue'
 import TokenRow from './TokenRow'
