+++ packages/lib/modules/tokens/token.types.ts
@@ -1,4 +1,4 @@
-import { GqlToken } from '@repo/lib/shared/services/api/generated/graphql'
+import { GetTokensQuery, GqlToken } from '@repo/lib/shared/services/api/generated/graphql'
 import { Address, HumanAmount } from '@balancer/sdk'
 
 export type TokenBase = Pick<GqlToken, 'address' | 'name' | 'symbol' | 'decimals' | 'chainId'>
@@ -43,4 +43,9 @@ export type BalanceMap = Map<string, AmountHumanReadable>
 
 export interface AmountHumanReadableMap {
   [address: string]: AmountHumanReadable
+}
+
+export type ApiToken = Omit<GetTokensQuery['tokens'][0], '__typename'> & {
+  underlyingToken?: ApiToken
+  weight?: string
 }
